backend/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Environment variables
python-dotenv==1.0.1
pydantic-settings==2.6.0

# PostgreSQL drivers (choose one or keep both)
psycopg2-binary==2.9.9   # sync
asyncpg==0.30.0          # async

# Scheduler
apscheduler==3.10.4

# File handling (Starlette handles most, but python-multipart is required)
python-multipart==0.0.9

# Hashing
passlib[bcrypt]==1.7.4
